Add files via upload
main.axf has interrupts enabled for the KEYs as well as an Interrupt manager called pushbutton_ISR that just toggles a value. For now the main program just draws an orange block in the middle of the screen. I plan on having the interrupt routine increment the block's speed and for the main program to have it move based on the speed set by the subroutine. The main program will probable also decelerate the block, perhaps through the use of a timer interrupt service routine. Main.axf is ready to load on cpulator.